Data tests

+++ domain/src/main/java/openweather/domain/models/Hourly.kt
@@ -9,7 +9,7 @@ data class Hourly (
     var pressure : Int?,
     var humidity : Int?,
     var dewPoint : Double?,
-    var uvi : Int?,
+    var uvi : Double?,
     var clouds : Int?,
     var visibility : Int?,
     var windSpeed : Double?,
